+++ bren/inc/zbmp.h
@@ -16,16 +16,16 @@
 // ZBMP on file
 struct ZBMPF
 {
-    short bo;
-    short osk;
-    short xpLeft;
-    short ypTop;
-    short dxp;
-    short dyp;
+    int16_t bo;
+    int16_t osk;
+    int16_t xpLeft;
+    int16_t ypTop;
+    int16_t dxp;
+    int16_t dyp;
     // void *rgb; 		// pixels follow immediately after ZBMPF
 };
 VERIFY_STRUCT_SIZE(ZBMPF, 12);
-const ulong kbomZbmpf = 0x55500000;
+const uint32_t kbomZbmpf = 0x55500000;
 
 /****************************************
     ZBMP class
@@ -40,33 +40,33 @@ class ZBMP : public ZBMP_PAR
     MARKMEM
 
   protected:
-    RC _rc;      // bounding rectangle of ZBMP
-    long _cbRow; // bytes per row
-    long _cb;    // count of bytes in Z buffer
-    byte *_prgb; // Z buffer
+    RC _rc;         // bounding rectangle of ZBMP
+    int32_t _cbRow; // bytes per row
+    int32_t _cb;    // count of bytes in Z buffer
+    uint8_t *_prgb; // Z buffer
     ZBMP(void)
     {
     }
 
   public:
-    static PZBMP PzbmpNew(long dxp, long dyp);
+    static PZBMP PzbmpNew(int32_t dxp, int32_t dyp);
     static PZBMP PzbmpNewFromBpmp(BPMP *pbpmp);
     static PZBMP PzbmpRead(PBLCK pblck);
-    static bool FReadZbmp(PCRF pcrf, CTG ctg, CNO cno, PBLCK pblck, PBACO *ppbaco, long *pcb);
+    static bool FReadZbmp(PCRF pcrf, CTG ctg, CNO cno, PBLCK pblck, PBACO *ppbaco, int32_t *pcb);
     ~ZBMP(void);
 
-    byte *Prgb(void)
+    uint8_t *Prgb(void)
     {
         return _prgb;
     }
-    long CbRow(void)
+    int32_t CbRow(void)
     {
         return _cbRow;
     }
 
-    void Draw(byte *prgbPixels, long cbRow, long dyp, long xpRef, long ypRef, RC *prcClip = pvNil,
+    void Draw(uint8_t *prgbPixels, int32_t cbRow, int32_t dyp, int32_t xpRef, int32_t ypRef, RC *prcClip = pvNil,
               PREGN pregnClip = pvNil);
-    void DrawHalf(byte *prgbPixels, long cbRow, long dyp, long xpRef, long ypRef, RC *prcClip = pvNil,
+    void DrawHalf(uint8_t *prgbPixels, int32_t cbRow, int32_t dyp, int32_t xpRef, int32_t ypRef, RC *prcClip = pvNil,
                   PREGN pregnClip = pvNil);
 
     bool FWrite(PCFL pcfl, CTG ctg, CNO *pcno);
